backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
obsws-python==1.5.1
moviepy==1.0.3
opencv-python==4.8.1.78
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
Pillow==10.2.0
